Solucionar problemas de superposición de atajos en VCUT. Agregar 'Acerca de'. Solucionar problemas 'duration' para yt-dlp. Ajustar tamaño de título de herramientas

+++ VManager/ViewModels/Herramienta5ViewModel.cs
@@ -192,7 +192,7 @@ namespace VManager.ViewModels
             }
         }
 
-        private string FormatDuration(int seconds)
+        private string FormatDuration(double seconds)
         {
             var ts = TimeSpan.FromSeconds(seconds);
             if (ts.TotalHours >= 1)
